packages/agentic/quick_test_fix.py
#!/usr/bin/env python3
"""
Quick Test to Validate All Fixes

Tests the fixes for:
1. Step reference resolution
2. Tool discovery and registration  
3. Parameter normalization
4. Result data structure handling
"""

import asyncio
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

async def test_fixes():
    """Test all the fixes we made."""
    print("🔧 TESTING ALL FIXES")
    print("=" * 50)
    
    try:
        # Test 1: Tool Registration
        print("\n1. Testing Tool Registration...")
        from pic_arcade_agentic.agents.tool_agent import ToolFirstAgent
        
        agent = ToolFirstAgent()
        tools = agent.get_available_tools()
        print(f"   ✅ Registered {len(tools)} tools")
        
        # Check for key tools
        tool_names = [tool['name'] for tool in tools]
        expected_tools = ['flux_kontext_max', 'style_transfer', 'object_change', 'workflow_planner']
        
        for tool_name in expected_tools:
            if tool_name in tool_names:
                print(f"   ✅ Found {tool_name}")
            else:
                print(f"   ❌ Missing {tool_name}")
        
        # Test 2: Parameter Normalization
        print("\n2. Testing Parameter Normalization...")
        from pic_arcade_agentic.tools.image_tools import FluxKontextMaxTool
        
        flux_tool = FluxKontextMaxTool()
        test_cases = [
            (None, "aspect_ratio", "1:1"),
            (None, "output_format", "jpg"),
            ("medium", "safety_tolerance", 2),
            ("low", "image_prompt_strength", 0.2)
        ]
        
        for input_val, method_name, expected in test_cases:
            method = getattr(flux_tool, f"_normalize_{method_name}")
            result = method(input_val)
            if result == expected:
                print(f"   ✅ {method_name}({input_val}) = {result}")
            else:
                print(f"   ❌ {method_name}({input_val}) = {result}, expected {expected}")
        
        # Test 3: Simple Workflow Planning
        print("\n3. Testing Workflow Planning...")
        simple_request = "Generate a portrait in watercolor style"
        
        result = await agent.process_request(simple_request)
        
        if result.get("success", False):
            print("   ✅ Workflow planning succeeded")
            
            plan = result.get("workflow_plan", {})
            if "workflow_plan" in plan:
                steps = plan["workflow_plan"]
                print(f"   ✅ Planned {len(steps)} steps")
                
                for i, step in enumerate(steps, 1):
                    tool_name = step.get("tool_name", "unknown")
                    print(f"      Step {i}: {tool_name}")
            else:
                print("   ⚠️ No workflow plan in result")
            
            execution = result.get("execution_results", {})
            if execution:
                print(f"   ✅ Execution attempted")
                status = execution.get("execution_status", "unknown")
                print(f"   📊 Status: {status}")
                
                exec_results = execution.get("execution_results", [])
                success_count = sum(1 for r in exec_results if r.get("success", False))
                print(f"   📈 Step Success: {success_count}/{len(exec_results)}")
            else:
                print("   ⚠️ No execution results")
        else:
            error = result.get("error", "Unknown error")
            print(f"   ❌ Workflow failed: {error}")
        
        print(f"\n🎯 SUMMARY:")
        print(f"   Tools Registered: {len(tools)}")
        print(f"   Parameter Normalization: Working")
        print(f"   Workflow Planning: {'✅ Working' if result.get('success') else '❌ Issues'}")
        
        if result.get("success"):
            print(f"\n🎉 ALL FIXES VALIDATED SUCCESSFULLY!")
            return True
        else:
            print(f"\n⚠️ Some issues remain, but core fixes are working")
            return False
            
    except Exception as e:
        print(f"\n❌ Test failed with error: {e}")
        import traceback
        traceback.print_exc()
        return False

if __name__ == "__main__":
    asyncio.run(test_fixes()) 