packages/agentic/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pic-arcade-agentic"
version = "0.1.0"
description = "Agentic AI backend for Pic Arcade using LangGraph"
authors = [
    {name = "Pic Arcade Team"}
]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "langgraph>=0.2.16",
    "langchain>=0.2.11", 
    "langchain-openai>=0.1.17",
    "langchain-community>=0.2.10",
    "openai>=1.40.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "httpx>=0.25.2",
    "replicate>=0.25.1",
    "pillow>=10.1.0",
    "numpy>=1.25.2",
    "requests>=2.31.0",
    "aiohttp>=3.9.1",
    "typing-extensions>=4.8.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88 